feat: overhauled the translation system
Translations are now stored separate from other quest data, under the
lang/ folder with the quest book folder hierarchy.
Editing language can be overridden via client preference; default is
to use whatever Minecraft language is in force.
Text which doesn't have a translation in the current locale (but does
in the en_us locale) is highlighted when in edit mode.