Prevents unnecessary text box update resulting in ui breakage
Avoids recreating the multiline text field when the width hasn't changed, which could cause text loss due to word wrapping.

This fixes the issue that's been occasionally reported and shown on video where the text wrapping breaks and the ui goes blank when typing on the first unwrapped lined of a description entry field and resolves itself on subsequent edits.

Resolves https://github.com/FTBTeam/FTB-Mods-Issues/issues/1330